+++ .changeset/breezy-snails-attend.md
@@ -1,0 +1,5 @@
+---
+'@redocly/cli': minor
+---
+
+Added the `@redocly/cli/api` entry point that exposes the `push` and `push-status` commands as the `push` and `pushStatus` functions.
