+++ packages/cli/package.json
@@ -8,6 +8,13 @@
     "openapi": "bin/cli.js"
   },
   "type": "module",
+  "exports": {
+    "./api": {
+      "types": "./lib/api.d.ts",
+      "default": "./lib/api.js"
+    },
+    "./package.json": "./package.json"
+  },
   "engines": {
     "node": ">=22.12.0 || >=20.19.0 <21.0.0",
     "npm": ">=10"
