+++ packages/cli/src/reunite/commands/__tests__/push-status.test.ts
@@ -1,6 +1,6 @@
+import { type PushResponse } from '../../../api.js';
 import * as errorHandling from '../../../utils/error.js';
 import { ReuniteApi } from '../../api/index.js';
-import { type PushResponse } from '../../api/types.js';
 import { handlePushStatus } from '../push-status.js';
 
 vi.mock('colorette', async () => {
