+++ packages/cli/src/reunite/commands/push-status.ts
@@ -1,58 +1,45 @@
 import { logger, type OutputFormat } from '@redocly/openapi-core';
 import * as colors from 'colorette';
 
+import type {
+  DeploymentStatus,
+  PushResponse,
+  PushStatusOptions,
+  PushStatusSummary,
+  ScorecardItem,
+} from '../../api.js';
 import type { VerifyConfigOptions } from '../../types.js';
 import { printExecutionTime, capitalize } from '../../utils/miscellaneous.js';
 import { Spinner } from '../../utils/spinner.js';
 import type { CommandArgs } from '../../wrapper.js';
 import { ReuniteApi, getApiKeys, getDomain } from '../api/index.js';
-import type {
-  DeploymentStatus,
-  DeploymentStatusResponse,
-  PushResponse,
-  ScorecardItem,
-} from '../api/types.js';
 import { DeploymentError } from '../utils.js';
 import { handleReuniteError, retryUntilConditionMet } from './utils.js';
 
 const RETRY_INTERVAL_MS = 5000; // 5 sec
 
-export type PushStatusArgv = {
-  organization: string;
-  project: string;
-  pushId: string;
-  domain?: string;
+export type PushStatusArgv = PushStatusOptions & {
   format?: Extract<OutputFormat, 'stylish'>;
-  wait?: boolean;
-  'max-execution-time'?: number; // in seconds
-  'retry-interval'?: number; // in seconds
-  'start-time'?: number; // in milliseconds
-  'continue-on-deploy-failures'?: boolean;
-  onRetry?: (lasSummary: PushStatusSummary) => void;
 } & VerifyConfigOptions;
 
-export interface PushStatusSummary {
-  preview: DeploymentStatusResponse;
-  production: DeploymentStatusResponse | null;
-  commit: PushResponse['commit'];
+export function handlePushStatus({ argv }: CommandArgs<PushStatusArgv>) {
+  return pushStatus(argv);
 }
 
-export async function handlePushStatus({
-  argv,
-}: CommandArgs<PushStatusArgv>): Promise<PushStatusSummary | void> {
+export async function pushStatus(options: PushStatusOptions): Promise<PushStatusSummary> {
   const startedAt = performance.now();
   const spinner = new Spinner();
 
-  const { organization, project: projectId, pushId, wait } = argv;
+  const { organization, project: projectId, pushId, wait } = options;
 
-  const domain = argv.domain || getDomain();
-  const maxExecutionTime = argv['max-execution-time'] || 1200; // 20 min
-  const retryIntervalMs = argv['retry-interval']
-    ? argv['retry-interval'] * 1000
+  const domain = options.domain || getDomain();
+  const maxExecutionTime = options['max-execution-time'] || 1200; // 20 min
+  const retryIntervalMs = options['retry-interval']
+    ? options['retry-interval'] * 1000
     : RETRY_INTERVAL_MS;
-  const startTime = argv['start-time'] || Date.now();
+  const startTime = options['start-time'] || Date.now();
   const retryTimeoutMs = maxExecutionTime * 1000;
-  const continueOnDeployFailures = argv['continue-on-deploy-failures'] || false;
+  const continueOnDeployFailures = options['continue-on-deploy-failures'] || false;
 
   try {
     const apiKey = getApiKeys();
@@ -82,8 +69,8 @@ export async function handlePushStatus({
         });
       },
       onRetry: (lastResult) => {
-        if (argv.onRetry) {
-          argv.onRetry({
+        if (options.onRetry) {
+          options.onRetry({
             preview: lastResult.status.preview,
             production: lastResult.isMainBranch ? lastResult.status.production : null,
             commit: lastResult.commit,
@@ -131,8 +118,8 @@ export async function handlePushStatus({
           });
         },
         onRetry: (lastResult) => {
-          if (argv.onRetry) {
-            argv.onRetry({
+          if (options.onRetry) {
+            options.onRetry({
               preview: lastResult.status.preview,
               production: lastResult.isMainBranch ? lastResult.status.production : null,
               commit: lastResult.commit,
@@ -169,7 +156,7 @@ export async function handlePushStatus({
   } catch (err) {
     spinner.stop(); // Spinner can block process exit, so we need to stop it explicitly.
 
-    handleReuniteError('✗ Failed to get push status.', err);
+    return handleReuniteError('✗ Failed to get push status.', err);
   } finally {
     spinner.stop(); // Spinner can block process exit, so we need to stop it explicitly.
   }
