+++ packages/cli/src/reunite/commands/push.ts
@@ -3,49 +3,32 @@ import { green } from 'colorette';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 
+import type { PushOptions, PushResult } from '../../api.js';
 import type { VerifyConfigOptions } from '../../types.js';
 import { exitWithError } from '../../utils/error.js';
 import { printExecutionTime } from '../../utils/miscellaneous.js';
-import { version } from '../../utils/package.js';
 import type { CommandArgs } from '../../wrapper.js';
 import { ReuniteApi, getDomain, getApiKeys } from '../api/index.js';
-import { handlePushStatus } from './push-status.js';
+import { pushStatus } from './push-status.js';
 import { handleReuniteError } from './utils.js';
 
-export type PushArgv = {
-  files: string[];
-  organization: string;
-  project: string;
-  'mount-path': string;
-  branch: string;
-  author: string;
-  message: string;
-  'commit-sha'?: string;
-  'commit-url'?: string;
-  namespace?: string;
-  repository?: string;
-  'created-at'?: string;
-  'default-branch': string;
-  domain?: string;
-  'wait-for-deployment'?: boolean;
-  'max-execution-time'?: number;
-  'continue-on-deploy-failures'?: boolean;
-  verbose?: boolean;
+export type PushArgv = PushOptions & {
   format?: Extract<OutputFormat, 'stylish'>;
 } & VerifyConfigOptions;
 
 type FileToUpload = { name: string; path: string };
 
-export async function handlePush({
-  argv,
-  config,
-}: CommandArgs<PushArgv>): Promise<{ pushId: string } | void> {
+export function handlePush({ argv }: CommandArgs<PushArgv>) {
+  return push(argv);
+}
+
+export async function push(options: PushOptions): Promise<PushResult | undefined> {
   const startedAt = performance.now(); // for printing execution time
   const startTime = Date.now(); // for push-status command
 
-  const { organization, project: projectId, 'mount-path': mountPath, verbose } = argv;
+  const { organization, project: projectId, 'mount-path': mountPath, verbose } = options;
 
-  const domain = argv.domain || getDomain();
+  const domain = options.domain || getDomain();
 
   if (!domain) {
     return exitWithError(
@@ -60,14 +43,15 @@ export async function handlePush({
       'default-branch': defaultBranch,
       'wait-for-deployment': waitForDeployment,
       'max-execution-time': maxExecutionTime,
-    } = argv;
-    const author = parseCommitAuthor(argv.author);
+    } = options;
+    const author = parseCommitAuthor(options.author);
     const apiKey = getApiKeys();
-    const filesToUpload = collectFilesToPush(argv.files);
+    const filesToUpload = collectFilesToPush(options.files);
     const commandName = 'push' as const;
 
     if (!filesToUpload.length) {
-      return printExecutionTime(commandName, startedAt, `No files to upload`);
+      printExecutionTime(commandName, startedAt, `No files to upload`);
+      return;
     }
 
     const client = new ReuniteApi({ domain, apiKey, command: commandName });
@@ -90,16 +74,16 @@ export async function handlePush({
       {
         remoteId: remote.id,
         commit: {
-          message: argv.message,
-          branchName: argv.branch,
+          message: options.message,
+          branchName: options.branch,
           sha: commitSha,
           url: commitUrl,
-          createdAt: argv['created-at'],
-          namespace: argv.namespace,
-          repository: argv.repository,
+          createdAt: options['created-at'],
+          namespace: options.namespace,
+          repository: options.repository,
           author,
         },
-        isMainBranch: defaultBranch === argv.branch,
+        isMainBranch: defaultBranch === options.branch,
       },
       filesToUpload.map((f) => ({ path: slash(f.name), stream: fs.createReadStream(f.path) }))
     );
@@ -114,19 +98,15 @@ export async function handlePush({
     if (waitForDeployment) {
       logger.info('\n');
 
-      await handlePushStatus({
-        argv: {
-          organization,
-          project: projectId,
-          pushId: id,
-          wait: true,
-          domain,
-          'max-execution-time': maxExecutionTime,
-          'start-time': startTime,
-          'continue-on-deploy-failures': argv['continue-on-deploy-failures'],
-        },
-        config,
-        version,
+      await pushStatus({
+        organization,
+        project: projectId,
+        pushId: id,
+        wait: true,
+        domain,
+        'max-execution-time': maxExecutionTime,
+        'start-time': startTime,
+        'continue-on-deploy-failures': options['continue-on-deploy-failures'],
       });
     }
     if (verbose) {
@@ -146,7 +126,7 @@ export async function handlePush({
       pushId: id,
     };
   } catch (err) {
-    handleReuniteError('✗ File upload failed.', err);
+    return handleReuniteError('✗ File upload failed.', err);
   }
 }
